@@ -2,29 +2,36 @@
   "name": "javascript-bridge",
   "version": "1.0.0",
   "description": "우아한테크코스 프리코스 다리 건너기 게임 미션",
+  "type": "module",
   "dependencies": {
-    "@woowacourse/mission-utils": "^1.0.6"
+    "@woowacourse/mission-utils": "2.1.1"
   },
   "devDependencies": {
-    "jest": "^29.0.3"
+    "@babel/core": "^7.23.0",
+    "@babel/preset-env": "^7.22.20",
+    "babel-jest": "^29.7.0",
+    "jest": "29.6.0"
   },
   "scripts": {
-    "test": "jest --detectOpenHandles --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/woowacourse/javascript-bridge.git"
+    "test": "jest --detectOpenHandles",
+    "start": "node ./src/index.js"
   },
   "keywords": [],
   "author": "woowacourse",
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/woowacourse/javascript-bridge/issues"
-  },
-  "homepage": "https://github.com/woowacourse/javascript-bridge#readme",
   "engineStrict": true,
+  "jest": {
+    "transform": {
+      "\\.js$": "babel-jest"
+    }
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
+  },
   "engines": {
-    "npm": ">=6.0.0",
-    "node": ">=14.0.0"
+    "npm": ">=9.6.7",
+    "node": ">=18.17.1"
   }
 }
