@@ -1,32 +1,36 @@
-const MissionUtils = require("@woowacourse/mission-utils");
-const App = require("../src/App");
-const BridgeMaker = require("../src/BridgeMaker");
+import { Console, Random } from '@woowacourse/mission-utils';
+import App from '../src/App.js';
+import BridgeMaker from '../src/BridgeMaker.js';
 
-const mockQuestions = (answers) => {
-  MissionUtils.Console.readLine = jest.fn();
-  answers.reduce((acc, input) => {
-    return acc.mockImplementationOnce((_, callback) => {
-      callback(input);
-    });
-  }, MissionUtils.Console.readLine);
+const mockQuestions = (inputs) => {
+  Console.readLineAsync = jest.fn();
+
+  Console.readLineAsync.mockImplementation(() => {
+    const input = inputs.shift();
+
+    if (input === undefined) {
+      throw new Error('NO INPUT');
+    }
+
+    return Promise.resolve(input);
+  });
 };
 
 const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
+  Random.pickNumberInRange = jest.fn();
+  numbers.reduce(
+    (acc, number) => acc.mockReturnValueOnce(number),
+    Random.pickNumberInRange,
+  );
 };
 
 const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, "print");
+  const logSpy = jest.spyOn(Console, 'print');
   logSpy.mockClear();
   return logSpy;
 };
 
-const getOutput = (logSpy) => {
-  return [...logSpy.mock.calls].join("");
-};
+const getOutput = (logSpy) => [...logSpy.mock.calls].join('');
 
 const runException = (inputs) => {
   mockQuestions(inputs);
@@ -35,7 +39,7 @@ const runException = (inputs) => {
 
   app.play();
 
-  expectLogContains(getOutput(logSpy), ["[ERROR]"]);
+  expectLogContains(getOutput(logSpy), ['[ERROR]']);
 };
 
 const expectLogContains = (received, logs) => {
@@ -51,37 +55,38 @@ const expectBridgeOrder = (received, upside, downside) => {
   expect(upsideIndex).toBeLessThan(downsideIndex);
 };
 
-describe("다리 건너기 테스트", () => {
-  test("다리 생성 테스트", () => {
-    const randomNumbers = ["1", "0", "0"];
-    const mockGenerator = randomNumbers.reduce((acc, number) => {
-      return acc.mockReturnValueOnce(number);
-    }, jest.fn());
+describe('다리 건너기 테스트', () => {
+  test('다리 생성 테스트', () => {
+    const randomNumbers = ['1', '0', '0'];
+    const mockGenerator = randomNumbers.reduce(
+      (acc, number) => acc.mockReturnValueOnce(number),
+      jest.fn(),
+    );
 
     const bridge = BridgeMaker.makeBridge(3, mockGenerator);
-    expect(bridge).toEqual(["U", "D", "D"]);
+    expect(bridge).toEqual(['U', 'D', 'D']);
   });
 
-  test("기능 테스트", () => {
+  test('기능 테스트', () => {
     const logSpy = getLogSpy();
-    mockRandoms(["1", "0", "1"]);
-    mockQuestions(["3", "U", "D", "U"]);
+    mockRandoms(['1', '0', '1']);
+    mockQuestions(['3', 'U', 'D', 'U']);
 
     const app = new App();
     app.play();
 
     const log = getOutput(logSpy);
     expectLogContains(log, [
-      "최종 게임 결과",
-      "[ O |   | O ]",
-      "[   | O |   ]",
-      "게임 성공 여부: 성공",
-      "총 시도한 횟수: 1",
+      '최종 게임 결과',
+      '[ O |   | O ]',
+      '[   | O |   ]',
+      '게임 성공 여부: 성공',
+      '총 시도한 횟수: 1',
     ]);
-    expectBridgeOrder(log, "[ O |   | O ]", "[   | O |   ]");
+    expectBridgeOrder(log, '[ O |   | O ]', '[   | O |   ]');
   });
 
-  test("예외 테스트", () => {
-    runException(["a"]);
+  test('예외 테스트', () => {
+    runException(['a']);
   });
 });
