@@ -1,7 +1,9 @@
+import { Console } from '@woowacourse/mission-utils';
+
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
  */
-const InputView = {
+export const InputView = {
   /**
    * 다리의 길이를 입력받는다.
    */
@@ -16,6 +18,9 @@ const InputView = {
    * 사용자가 게임을 다시 시도할지 종료할지 여부를 입력받는다.
    */
   readGameCommand() {},
-};
 
-module.exports = InputView;
+  async readUserInput(message) {
+    const input = await Console.readLineAsync(message);
+    return input;
+  },
+};
